==> 02_exposure_fltr.Rmd <==
---
title: "02_exposure_fltr"
author: "Puvvula"
date: "2024-06-07"
output: pdf_document
---

#exposure-linkage
```{r}
pt_id<- read_csv(paste0(dat_path, "01_research/PtReg.csv")) |> 
  select(c(1,8,9,11:15,17)) |> distinct()|>  clean_names()

exp_linkage<- read_csv(paste0(dat_path, "05_assays/Lab_Bio_Specimens.csv")) |>
  select(c(2,4,6,8,13,14)) |> distinct()|> clean_names()
```

#exposure data
```{r}
exp_dat<- read_csv(paste0(dat_path, "05_assays/Lab_Bio_Analysis_Results.csv")) |>
  select(-c(1:3,6,8:12,14,15,18,20:26)) |>
  mutate(analysis_analyte_name = str_to_lower(analysis_analyte_name),
         #log_lod = log2(analysis_llod + 1e-6)
         )|>
  mutate(analysis_analyte_name = fct_recode(analysis_analyte_name,
                                            "bis(butoxyethyl) phosphate" = "bis(butoxyethyl) pohosphate",
                                            "bis(2-ethylhexyl) phosphate" = "bis(2-ethylhexyl) phoshate",
                                            "bis(1,3-dichloro-2-propyl) phosphate" = "bis(1,3-dichloro-2propyl) phosphate",
                                            "bis(1-chloro-2-propyl) phosphate" = "bis-(1-chloro-2propl) phosphate",
                                            "bis(2-chloroethyl) phosphate" = "bis-2chloroethyl phosphate")) |>
  clean_names() |>
  filter(analysis_analyte_name %in% 
           c("dipropyl phosphate",
             "bis(2-ethylhexyl) phosphate",
             "bis(2-methylphenyl) phosphate",
             "bis(1-chloro-2-propyl) phosphate",
             "bis(2-chloroethyl) phosphate",
             "bis(butoxyethyl) phosphate",
             "bis(1,3-dichloro-2-propyl) phosphate",
             "diethylphosphate"))
```


#join exposure data with exposure linkage file
```{r}
exp_dat_df<- left_join(exp_dat, exp_linkage, by= "x_specimen_id") 
```

```{r}
summary_df <- exp_dat_df |>
  mutate(specimen_collection_life_stage = paste0("lf_st_", specimen_collection_life_stage),
         specimen_collection_trimester = paste0("trim_", specimen_collection_trimester)) |>
  group_by(analysis_analyte_name, specimen_collection_life_stage, specimen_collection_trimester) |>
  summarize(
    observation_count = n()
  ) |>
  pivot_wider(names_from = c(specimen_collection_life_stage, specimen_collection_trimester), 
              values_from = observation_count)
```
