==> 06_exposure_tbl_grant.Rmd <==
---
title: "06_exposure_tbl_grant"
author: "Puvvula"
date: "2024-12-10"
output: pdf_document
---

#exposure-linkage
```{r}
pt_id<- read_csv(paste0(dat_path, "01_research/PtReg.csv")) |> 
  select(c(4,1,8,11)) |> 
  #distinct()|>  
  clean_names() |>
  mutate(participant_type = fct_recode(participant_type,
                                     "preg_woman" = "P",
                                     "child" = "C")) |>
  filter(participant_type == "preg_woman")

exp_linkage <- read_csv(paste0(dat_path, "05_assays/Lab_Bio_Specimens.csv")) |>
    select(c(2, 4, 6, 8, 9, 13, 14)) |>
    clean_names() |>
    mutate(
        specimen_collection_life_stage = factor(specimen_collection_life_stage) |>
            fct_recode(
                "pre_conception" = "1",
                "prenatal" = "2",
                "birth_to_12month" = "3",
                "early_child" = "4",
                "middle_child" = "5",
                "adolescen" = "6",
                "adult" = "7",
                "NA" = "-6",
                "declined" = "-7",
                "dont_know" = "-8",
                "missing" = "-9"
            ),
        specimen_collection_trimester = factor(specimen_collection_trimester) |>
            fct_recode(
                "0_13Weeks" = "1",
                "14_26Weeks" = "2",
                "27Wk_delivery" = "3",
                "NA" = "-6",
                "declined" = "-7",
                "dont_know" = "-8",
                "missing" = "-9"
            ),
        specimen_urine_collect_type = factor(specimen_urine_collect_type) |>
            fct_recode(
                "spot" = "1",
                "morning_first_void" = "2",
                "24hr_smpl" = "3",
                "NA" = "-6",
                "declined" = "-7",
                "dont_know" = "-8",
                "missing" = "-9"
            )
    ) |>
  filter(specimen_collection_life_stage == "prenatal")
```

#biospecimen type
```{r}
specimen_type<- read_csv(paste0(dat_path, "05_assays/SpecimenTypes.csv")) |>
  clean_names() |>select(c(1,2))
```

#perform linkage
```{r}
exp_linkage_update <- left_join(exp_linkage, pt_id, by= "x_participant_id") |>
  select(c(1:6,7,9)) |>
  distinct() |>
  mutate(
    specimen_type_id = case_when(
      specimen_urine_collect_type %in% c("spot", "morning_first_void", "24_hr_smpl") ~ "Urine",
      TRUE ~ specimen_type$specimen_type[match(specimen_type_id, specimen_type$specimen_type_id)]
    )
  )
```


#exposure data
```{r}
exp_dat<- read_csv(paste0(dat_path, "05_assays/Lab_Bio_Analysis_Results.csv")) |>
  select(-c(1:3,6,8:12,15,18,20:26)) |>
  mutate(analysis_analyte_name = str_to_lower(analysis_analyte_name),
         #log_lod = log2(analysis_llod + 1e-6)
         )|>
  mutate(analysis_analyte_name = fct_recode(analysis_analyte_name,
                                            "bis(butoxyethyl) phosphate" = "bis(butoxyethyl) pohosphate",
                                            "bis(2-ethylhexyl) phosphate" = "bis(2-ethylhexyl) phoshate",
                                            "bis(1,3-dichloro-2-propyl) phosphate" = "bis(1,3-dichloro-2propyl) phosphate",
                                            "bis(1-chloro-2-propyl) phosphate" = "bis-(1-chloro-2propl) phosphate",
                                            "bis(2-chloroethyl) phosphate" = "bis-2chloroethyl phosphate"),
         analysis_analyte_form = factor(analysis_analyte_form) |>
           fct_recode(
                                            "free" = "1",
                                            "bound" = "2",
                                            "total" = "3",
                                            "speciated" = "4",
                                            "other" = "5",
                                            "not_applic" = "-6",
                                            "declined" = "-7",
                                            "dont_know" = "-8",
                                            "missing" = "-9")
         ) |>
  clean_names() |>
  filter(x_specimen_id %in% exp_linkage_update$x_specimen_id)
```

#biomarker dictonary
```{r}
unique_table <- exp_dat |> 
  distinct(analyte_id, analysis_analyte_name) |> 
  group_by(analyte_id) |> 
  summarise(analysis_analyte_names = paste(analysis_analyte_name, collapse = ", "), .groups = "drop") |>
  mutate(
    biomarker_name = str_replace_all(
      str_split(analysis_analyte_names, ", ") |> 
        sapply(\(x) x[1]),  # Extract the first value after split
      "[^[:alnum:]]+",      # Replace non-alphanumeric characters
      "_"
    )
  ) |>
  select(c(1,3))

unique_table_org<- read_csv(paste0(dat_path, "05_assays/Analytes.csv")) |>
  clean_names() |>
  select(1:3,5) |>
  mutate(class_name = case_when(
    assay_class_id == "C007" ~ "Per- and polyfluoroalkyl substances",
    assay_class_id == "C010" ~ "Phthalates",
    assay_class_id == "C013" ~ "Phenols and Parabens",
    assay_class_id == "C016" ~ "OPEs",
    assay_class_id == "C019" ~ "Metals",
    assay_class_id == "C022" ~ "Cotinine related",
    assay_class_id == "C025" ~ "Agri_chemicals_p1",
    assay_class_id == "C031" ~ "Agri_chemicals_p2",
    assay_class_id == "C037" ~ "PAHs",
    assay_class_id == "C049" ~ "Vit-D",
    assay_class_id == "C052" ~ "Organic_compounds_aniline",
    assay_class_id == "C055" ~ "Agri_chemicals_p3",
    assay_class_id == "C058" ~ "Agri_chemicals_p4",
    assay_class_id == "C061" ~ "Agri_chemicals_p5",
    TRUE ~ as.character(assay_class_id)  # For unmatched cases
  ))
```


#join exposure data with exposure linkage file
```{r}
exp_dat_df<- left_join(exp_dat, exp_linkage_update, by= "x_specimen_id")  |>
  left_join(unique_table_org, by = "analyte_id") |>
  mutate(specimen_combined = paste(specimen_urine_collect_type, specimen_type_id, sep = "_")) |>
  select(c(9,10,2,1,3,14,5,16:20,6:8)) |>
  clean_names()|>
  mutate(specimen_combined = recode(specimen_combined,
                                    "morning_first_void_Urine" = "urine",
                                    "spot_Urine" = "urine",
                                    "24hr_smpl_Not otherwise specified" = "urine",
                                    "NA_Plasma" = "plasma",
                                    "NA_Serum" = "serum",
                                    "NA_Red blood cells (RBCs)" = "rbc",
                                    "NA_Whole" = "whole_blood",
                                    "dont_know_Not otherwise specified" = "NA",
                                    "missing_Not otherwise specified" = "NA",
                                    "NA_Not otherwise specified" = "NA")) |>
  group_by(analyte_abbrev) |>
  mutate(
  specimen_combined = if (any(specimen_combined == "urine", na.rm = TRUE)) {
    specimen_combined  # Keep "urine" as separate level
  } else {
    paste(unique(specimen_combined), collapse = "_")  # Combine if "urine" is not present
  }
)|>
  mutate(specimen_collection_trimester = case_when(
    specimen_collection_trimester %in% c("missing", "dont_know", "declined", "NA", " ") ~ "Missing",
    is.na(specimen_collection_trimester) ~ "Missing",
    TRUE ~ specimen_collection_trimester
  )) |>
  ungroup()

#get unique list to filter biomarkers
#split(exp_dat_df$analyte_abbrev, exp_dat_df$class_name) |> lapply(unique)

#filter analytes mentioend in Aimin's grant
skip_biomarker_filter <- TRUE

filtered_exp_df <- exp_dat_df |> 
  (\(data) {
    if (skip_biomarker_filter) {
      data
    } else {
      data |> filter(analyte_abbrev %in% 
        c(
          "Pb", "Hg-Total", "MeHg", "Cd", "As", "Ni", "V",
          "DMDP or DMDTP", "DMP", "DMTP", "DEP", "DEDP or DEDTP",
          "PFOS", "PFOA", "PFHXS", "PFNA", 
          "BPAP", "BPAF", "BPA",
          "MEP", "MNBP", "MEHP", "MEHHP", "MEOHP", "MECPP", 
          "TCS", 
          "BuPB", "MEPB", "PRPB", "ETPB", 
          "BDCPP", "DPhP", "BCEtP", "DBuP", "specific gravity"
        ))
    }
  })() |> 
  filter(class_name %in% 
    c(
      "Agri_chemicals_p1", "Agri_chemicals_p2", "Agri_chemicals_p3", "Agri_chemicals_p4", "Agri_chemicals_p5",
      "Metals", "OPEs", "Organic_compounds_aniline", "PAHs", "Per- and polyfluoroalkyl substances",
      "Phenols and Parabens", "Phthalates", "C001"
    )
  ) |>
  mutate(analysis_result_unit = case_when(
    analysis_result_unit %in% c("ng/ml", "ng/Ml") ~ "ng/mL",
    analysis_result_unit == "ug/L" ~ "ng/mL",
    analysis_result_unit == "ppb" ~ "ng/mL",
    analysis_result_unit == "ug/dl" ~ "ug/dL",
    TRUE ~ analysis_result_unit
  ))

#added urine specific gravity variable on Feb 19, 2025
#save(filtered_exp_df, file="~/Documents/ECHO/grant_draft/ECHObiomarkers_for_grant_feb192025.rda")
```

```{r}
analysis_analyte_counts <- filtered_exp_df |>
  group_by(assay_class_id, class_name, specimen_collection_trimester, 
           analyte_name, analyte_abbrev, analysis_result_unit,specimen_combined) |>
  summarise(
    analyte_count = n(),
    less_than_llod_count = sum(analysis_result < analysis_llod, na.rm = TRUE),
    .groups = "drop"
  ) |>
  mutate(
    less_than_llod_percent = (less_than_llod_count / analyte_count) * 100,
    lt_lod_cnt_and_pct = sprintf("%d (%.2f%%)", less_than_llod_count, less_than_llod_percent)
  ) |>
  select(-c(less_than_llod_count, less_than_llod_percent)) |>
  pivot_wider(
    names_from = specimen_collection_trimester,
    values_from = c(analyte_count, lt_lod_cnt_and_pct),
    values_fill = list(analyte_count = 0, lt_lod_cnt_and_pct = "0 (0.00%)")
  ) |>
  mutate(
    Total_unique_participants = map2_dbl(
      analyte_name, specimen_combined,
      ~ n_distinct(exp_dat_df$x_participant_id[
        exp_dat_df$analyte_name == .x & exp_dat_df$specimen_combined == .y
      ])
    )
  ) |>
  filter(
    !(class_name %in% c("Phenols and Parabens") & 
      grepl("serum|plasma", specimen_combined, ignore.case = TRUE))
  ) |>
  filter(
    !(class_name %in% c("Phthalates") & 
      grepl("serum|plasma", specimen_combined, ignore.case = TRUE))
  ) |>
  filter(Total_unique_participants >300)

#save(analysis_analyte_counts, file="~/Documents/ECHO/prelim_results/exposure_counts_dec_23.rda")

#write_csv(analysis_analyte_counts, "~/Documents/ECHO/prelim_results/exposure_counts.csv")

#write_csv(analysis_analyte_counts, "~/Documents/ECHO/grant_draft/ECHObiomarkers.csv")
```

#for viz only
```{r}
analysis_analyte_counts_long <- analysis_analyte_counts %>%
  pivot_longer(
    cols = c(4:8), # This selects all columns from the 4th one onward
    names_to = "time_point", # New column for the time point
    values_to = "count" # New column for the values
  )

ggplot(analysis_analyte_counts_long, aes(x = class_name, y = count, fill = analyte_name)) +
  geom_bar(stat = "identity") +
  theme_minimal() +
  labs(title = "", x = "Class Name", y = "Count") +
  theme(
    legend.position = "none",  # Remove the fill legend
    axis.text.x = element_text(angle = 45, hjust = 1)  # Rotate x-axis labels to 45 degrees
  ) +
  facet_grid(time_point~.)
```

#for geometric mean - imputed <LOD with LOD/sqrt-2
```{r}
gm_df<- filtered_exp_df |>
  mutate(
    analysis_result = if_else(
      analysis_result < analysis_llod, 
      analysis_llod / sqrt(2), 
      analysis_result
    ),
    analysis_result = round(analysis_result, 3) 
  )

#geometric mean function
gm_mean = function(x, na.rm=TRUE){
    exp(sum(log(x[x > 0]), na.rm=na.rm) / length(x))
}

gm_summ <- gm_df |>
  group_by(class_name, analyte_name, analyte_abbrev, specimen_combined, analysis_result_unit, specimen_collection_trimester) |> 
  summarise(
    count = n(),
    geometric_mean = sprintf("%.2f", gm_mean(analysis_result)), # 
    median = sprintf("%.2f", median(analysis_result, na.rm = TRUE)),
    .groups = "drop"
  ) |>
  filter(specimen_collection_trimester == "14_26Weeks")

write_csv(gm_summ, "~/Documents/ECHO/grant_draft/ECHObiomarkersGEOMETRICMean_jan14.csv")
```

#join lod counts and geometric mean/median from second trimester
```{r}
summary_df<- analysis_analyte_counts |>
  left_join(gm_summ, c("analyte_name", "specimen_combined", "analysis_result_unit")) |>
  select(c(1:15,20,21)) |>
  rename(geometric_mean_2trim = geometric_mean,
         median_2trim = median) |>
  clean_names() |>
  select(-c(median_2trim))

write_csv(summary_df, "~/Documents/ECHO/grant_draft/ECHO_lod_count_geoMean.csv")423
```


#probability density plot of analytes
```{r}
plot_df<- gm_df |>
  filter(specimen_collection_trimester == "14_26Weeks",
         class_name == "Metals")

ggplot(plot_df, aes(x=analysis_result, color=specimen_combined))+
  geom_density()+
  theme_minimal()+
  facet_wrap(.~analyte_name, scale="free")
  theme(
    legend.position = "none"
  )
```
